+++ lib/legion/extensions/transformer/transport.rb
@@ -4,7 +4,11 @@ module Legion
   module Extensions
     module Transformer
       module Transport
-        extend Legion::Extensions::Transport
+        def self.build
+          extend ::Legion::Extensions::Transport unless @_extended
+          @_extended = true
+          super
+        end
 
         def self.additional_e_to_q
           [